% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.r
\name{add_package}
\alias{add_package}
\title{add_package}
\usage{
add_package(pkg, root = "/hpcran/", emails = NULL)
}
\arguments{
\item{pkg}{File path of package in staging area.}

\item{root}{hpcran root path}

\item{emails}{list of verified emails}
}
\description{
Add/update package to the archive.
}
