% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenerate.r
\name{regenerate}
\alias{regenerate}
\title{regenerate}
\usage{
regenerate(root = "/hpcran/")
}
\arguments{
\item{root}{hpcran root path}
}
\description{
Regenerate the package html, RAN-files, and archive paths.
}
